pyproject.toml
[project]
name = "english-podcast"
version = "0.1.0"
description = ""
authors = [
    {name = "forest6511",email = "20209757+forest6511@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "moviepy (>=2.2.1,<3.0.0)",
    "numpy (>=2.2.6,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "google-cloud-texttospeech (>=2.27.0,<3.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "rich (>=14.0.0,<15.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "scipy (>=1.15.3,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "imageio-ffmpeg (>=0.6.0,<0.7.0)",
    "click (>=8.2.1,<9.0.0)",
    "pydub (>=0.25.1,<0.26.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"